knowledge-base/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
Authlib==1.6.4
blinker==1.9.0
cachetools==6.2.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==45.0.6
deprecation==2.1.0
distro==1.9.0
Django==5.2.5
django-allauth==65.11.1
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
grpcio==1.75.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
MarkupSafe==3.0.2
ml_dtypes==0.5.3
numpy==2.3.3
oauthlib==3.3.1
onnx==1.19.0
openai==1.108.1
openpyxl==3.1.5
packaging==25.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pprintpp==0.4.0
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
PyMuPDF==1.26.4
pyparsing==3.2.5
pypdfium2==4.30.0
python-dotenv==1.1.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
simplejwt==2.0.1
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing==3.7.4.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
validators==0.35.0
weaviate-client==4.17.0
Werkzeug==3.1.3
